+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - Add `lunaris._onInvalidate(event, handler)` to capture events in invalidation module. Only `invalidate` event is sent. It allows an handler to capture invalidations when offline mode is activated.
   - The offline app is not anymore mount when the the browser is online. The `lunaris._vue.mountOfflineApp` function must be used.
   - Add `lunaris.load(store)` function. It loads data of a store. The flag `lunaris.offline.isOfflineMode` must equal `true`.
+  - Add event `insert` for store `lunarisOfflineTransaction` (i.e. where HTTP transactions are stored in offline/offline mode).
 - Fix:
 
 ## 1.12.0